python/requirements.txt
# requirements
argon2-cffi==21.3.0
argon2.cffi-bindings==21.2.0
bcrypt==4.0.1
cffi==1.15.1
construct==2.10.68
cryptography==39.0.2
fabric==3.2.2 # previous used 3.0.0
future==0.18.3
invoke==2.2.0 # previous used 2.0.0
lxml==5.1.0 # previous used 4.9.2
markdown-it-py==2.2.0
mdurl==0.1.2
paramiko==3.1.0
pycparser==2.21
pycryptodomex==3.17
Pygments==2.14.0
pykeepass==4.0.3
PyNaCl==1.5.0
python-dateutil==2.8.2
rich==13.3.2
six==1.16.0
#
# additional requirements
pika==1.3.2
eciespy==0.4.2